# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eletro AR - Admin",
        "version": "0.1"
    },
    "paths": {
...
        "/api/financial": {
            "get": {
                "tags": [
                    "Financial"
                ],
                "summary": "Listar operacoes financeiras",
                "description": "Retorna uma lista de operacoes financeiras",
                "operationId": "a730ab76f18e2c3531dea7d0dd837677",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página a ser recuperada",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "financial": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "value": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "type": "boolean"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Financial"
                ],
                "summary": "Criar uma operacao",
                "description": "Cria uma nova operacao financeira",
                "operationId": "b747989b01f5d0e0b43a1310117f0d61",
                "requestBody": {
                    "description": "Dados da operacao",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "description",
                                    "value",
                                    "type"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ordem de serviço criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "service_order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/financial/{financial}": {
            "get": {
                "tags": [
                    "Financial"
                ],
                "summary": "Detalhes da operacao",
                "description": "Retorna os detalhes da operacao ",
                "operationId": "31ff56c8319507058a2a83e18193bbb3",
                "parameters": [
                    {
                        "name": "financial",
                        "in": "path",
                        "description": "ID da operacao financeira",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "financial": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operacao não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Financial"
                ],
                "summary": "Atualizar operacao financeira",
                "description": "Atualiza uma operacao financeira",
                "operationId": "548c30878f19285b3f0a4603103c1144",
                "parameters": [
                    {
                        "name": "financial",
                        "in": "path",
                        "description": "ID da operacao financeira",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da operacao",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "description",
                                    "value",
                                    "type"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ordem de serviço atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "financial": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Financial"
                ],
                "summary": "Deletar operacao financeira",
                "description": "Deletar uma operacao financeira",
                "operationId": "ad9bf691652f0487ac5563401f0c4f61",
                "parameters": [
                    {
                        "name": "financial",
                        "in": "path",
                        "description": "ID da operacao financeira",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacao financeira deletada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operacao financeira não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}